+++ cvmfs/catalog_mgr.h
@@ -139,7 +139,8 @@ template <class CatalogT>
 class AbstractCatalogManager : public SingleCopy {
  public:
   typedef std::vector<CatalogT*> CatalogList;
-
+  typedef CatalogT catalog_t;
+  
   static const inode_t kInodeOffset = 255;
   explicit AbstractCatalogManager(perf::Statistics *statistics);
   virtual ~AbstractCatalogManager();
